Better file structure for auth
