+++ backend/apps/core/authentications.py
@@ -35,25 +35,8 @@ from typing import Optional, Tuple
 from django.conf import settings
 from rest_framework.authentication import BaseAuthentication, get_authorization_header
 from rest_framework.exceptions import AuthenticationFailed
-from drf_spectacular.extensions import OpenApiAuthenticationExtension
 
 
-class ApiKeyAuthenticationScheme(OpenApiAuthenticationExtension):
-    """
-    OpenAPI extension for drf-spectacular to document Api-Key auth.
-    """
-    target_class = 'apps.core.authentications.ApiKeyAuthentication'
-    name = 'ApiKey'
-    match_subclasses = True
-
-    def get_security_definition(self, auto_schema):
-        return {
-            'type': 'apiKey',
-            'in': 'header',
-            'name': 'Authorization',
-            'description': 'Use `Api-Key <KEY>` to authenticate.',
-        }
-    
 class ApiKeyAuthentication(BaseAuthentication):
     """
     DRF authentication class that validates ``Api-Key`` request headers.
